Ensure that strings used for string kernel functions only consist of ACGT

+++ R/kernels.R
@@ -124,7 +124,9 @@ coerceStringInput <- function(x, k.info, ...) {
 
   sgg('add_preproc', k.info$preproc$type)
   if (target == "TRAIN") {
-    sgg(add.kernel, k.info$static, 'WORD', cache, params$use.sign,
+    # sgg(add.kernel, k.info$static, 'WORD', cache, params$use.sign,
+    #     params$normalization)
+    sgg(add.kernel, k.info$static, 'ULONG', cache, params$use.sign,
         params$normalization)
   }
